Separate classes into specific files
