+++ requirements.txt
@@ -3,7 +3,6 @@ isort==5.5.4
 lazy-object-proxy==1.4.3
 mccabe==0.6.1
 pygame==2.0.0.dev12
-pylint==2.6.0
 rope==0.18.0
 six==1.15.0
 toml==0.10.1
