+++ snake.py
@@ -1,16 +1,10 @@
 import pygame
 from pygame.locals import *
-import random
 
 GREEN = (0, 255, 0)
 CYAN = (0, 255, 255)
-RED = (255, 0, 0)
-
-# Initialise pygame
-pygame.init()
 
 
-# SnakeSegment and Snake classes
 class SnakeSegment(pygame.sprite.Sprite):
 
     def __init__(self, size, pos_x, pos_y, colour=GREEN):
@@ -107,77 +101,3 @@ class Snake():
 
     def draw(self, surface):
         self.all_segments.draw(surface)
-
-
-# Apple class
-class Apple(pygame.sprite.Sprite):
-
-    def __init__(self, max_x, max_y):
-        super().__init__()
-        self.size = 10
-        self.image = pygame.surface.Surface((self.size, self.size))
-        self.image.fill(RED)
-        self.pos_x = random.randint(0, max_x)
-        self.pos_y = random.randint(0, max_y)
-        self.rect = self.image.get_rect(center=(self.pos_x, self.pos_y))
-
-
-# Main game loop function
-def game_loop():
-    pygame.init()
-    clock = pygame.time.Clock()
-
-    screen_width = 500
-    screen_height = 500
-    screen = pygame.display.set_mode((screen_width, screen_height))
-
-    running = True
-
-    snake = Snake(screen_width / 2, screen_height / 2)
-    apple = Apple(screen_width, screen_height)
-
-    collided = False
-
-    while running:
-
-        clock.tick(60)
-
-        for event in pygame.event.get():
-            if event.type == QUIT:
-                running = False
-
-        screen.fill((0, 0, 0))
-
-        if apple is None:
-            apple = Apple(screen_width, screen_height)
-
-        screen.blit(apple.image, apple.rect)
-
-        if pygame.sprite.collide_rect(snake.head, apple):
-            snake.grow()
-            apple = None
-
-        pressed_keys = pygame.key.get_pressed()
-
-        tail_collision = pygame.sprite.spritecollideany(
-            snake.head, snake.tail_sprites)
-        if tail_collision:
-            collided = True
-            running = False
-
-        snake.update(pressed_keys, screen_width, screen_height)
-        snake.draw(screen)
-
-        pygame.display.update()
-
-    return collided
-
-
-# Run the game
-collided = game_loop()
-
-if collided:
-    pygame.time.wait(5000)
-
-# Shut down pygame once the game loop ends
-pygame.quit()
